Pull request 768: Update project layout to agent-friendly
Squashed commit of the following:

commit d0166bc094a4f688de049ecbbc40031aa9405e13
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Thu Apr 9 14:11:56 2026 +0300

    fix by commets

commit 32a58c6f631fe9d423357c75c220cfe216fe9a91
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Tue Apr 7 21:46:29 2026 +0300

    more fix clang
    fix win hash

commit 64be86598260539f3022a6d49aa1d315e8c948b3
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Tue Apr 7 21:01:59 2026 +0300

    add include and fix

commit 87011314cc63dea0cfed591f19941ff24d0aa2df
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Tue Apr 7 20:28:21 2026 +0300

    fix check LLVM 21

commit 6c348352ce99d370bd09185927869eebe254751d
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Tue Apr 7 20:24:03 2026 +0300

    fix windows

commit 1021955f417f0f91c3b7d2afe3ff421e4f2a4d6a
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Tue Apr 7 20:14:05 2026 +0300

    small fix and add check clang-format

commit ef1eaa2e2b7347062907dd4c5b80b1c4d6609581
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Tue Apr 7 19:56:49 2026 +0300

    Fixing current clang-format/clang-tidy errors: part 3

commit 95c0d3fb3aee0c9bb446c7a3fd6e3b47a71a5804
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Tue Apr 7 19:44:32 2026 +0300

    Fixing current clang-format/clang-tidy errors: part 2

commit 19a86642f10a311db7b4440922cdd8fb4587d8c2
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Tue Apr 7 13:00:26 2026 +0300

    Fixing current clang-format/clang-tidy errors: part 1

commit b50cdb5e7a093f4f71a90c2d0f5331bbf621502e
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Fri Apr 3 17:10:19 2026 +0300

    copy clang-tidy from vpn-libs and run make lint-fix-cpp

commit 2a990f2c6c9a6343cd26eccafc87e911cdb5528c
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Fri Apr 3 14:12:11 2026 +0300

    refactoring agentmd

commit 054f2015dd9943ebae31859d4bb2f26cdde09c99
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Fri Apr 3 12:32:04 2026 +0300

    adaptation for dns-libs

commit 22acd717560993d949033b1bbb623b22c83b9208
Author: Yuriy Selivanov <y.selivanov@adguard.com>
Date:   Fri Apr 3 10:16:14 2026 +0300

    copied from vpnlibs

+++ dnsfilter/test/benchmark.cpp
@@ -7,8 +7,8 @@
 #include "common/clock.h"
 #include "common/file.h"
 #include "common/logger.h"
-#include "dns/common/sys.h"
 #include "common/utils.h"
+#include "dns/common/sys.h"
 #include "dns/dnsfilter/dnsfilter.h"
 
 #undef max // `Nanos::max()` conflicts with `max` macro from `minwindef.h` on Windows
@@ -61,14 +61,14 @@ typedef struct {
     } overall;
 } test_result_t;
 
-static constexpr std::string_view HELP_MESSAGE
-        = "DNS filter benchmarking utility\n"
-          "\n"
-          "Usage: dnsfilter_benchmark [options...]\n"
-          "\n"
-          "    -h           print this message\n"
-          "    -f <path>    path to filter list file (default='" DEFAULT_FILTER_PATH "')\n"
-          "    -d <path>    path to domains list file (default='" DEFAULT_DOMAINS_BASE_PATH "')\n";
+static constexpr std::string_view HELP_MESSAGE =
+        "DNS filter benchmarking utility\n"
+        "\n"
+        "Usage: dnsfilter_benchmark [options...]\n"
+        "\n"
+        "    -h           print this message\n"
+        "    -f <path>    path to filter list file (default='" DEFAULT_FILTER_PATH "')\n"
+        "    -d <path>    path to domains list file (default='" DEFAULT_DOMAINS_BASE_PATH "')\n";
 
 static std::vector<std::string> domains;
 
@@ -112,8 +112,8 @@ static int apply_filter_to_base(test_result_t *tr, DnsFilter *filter, DnsFilter:
         tr->match_domains.total_matches += rules.size();
 
         if (!effective_rules.leftovers.empty()) {
-            if (const auto *adblock_info
-                    = std::get_if<DnsFilter::AdblockRuleInfo>(&effective_rules.leftovers[0]->content);
+            if (const auto *adblock_info =
+                            std::get_if<DnsFilter::AdblockRuleInfo>(&effective_rules.leftovers[0]->content);
                     adblock_info != nullptr && !adblock_info->props.test(DnsFilter::DARP_EXCEPTION)) {
                 tr->match_domains.effective_blocking_matches += effective_rules.leftovers.size();
             } else {
@@ -146,8 +146,8 @@ static int apply_filter_to_base(test_result_t *tr, DnsFilter *filter, DnsFilter:
 static void report_results(const test_result_t *result) {
     infolog(g_log, "============================================");
     infolog(g_log, "Load rules measurements:");
-    std::chrono::duration elapsed
-            = std::chrono::duration<double, std::ratio<1>>(result->load_rules.end_ts - result->load_rules.start_ts);
+    std::chrono::duration elapsed =
+            std::chrono::duration<double, std::ratio<1>>(result->load_rules.end_ts - result->load_rules.start_ts);
     infolog(g_log, "\tTime elapsed:               {}s", elapsed.count());
     infolog(g_log, "\tRSS before:                 {}kB", result->load_rules.start_rss);
     infolog(g_log, "\tRSS after:                  {}kB", result->load_rules.end_rss);
@@ -157,8 +157,8 @@ static void report_results(const test_result_t *result) {
     infolog(g_log, "\tTotal rules matched:        {}", result->match_domains.total_matches);
     infolog(g_log, "\tEffective blocking rules:   {}", result->match_domains.effective_blocking_matches);
     infolog(g_log, "\tEffective exception rules:  {}", result->match_domains.effective_exception_matches);
-    elapsed = std::chrono::duration<double, std::ratio<1>>(
-            result->match_domains.end_ts - result->match_domains.start_ts);
+    elapsed =
+            std::chrono::duration<double, std::ratio<1>>(result->match_domains.end_ts - result->match_domains.start_ts);
     infolog(g_log, "\tTime elapsed:               {}s", elapsed.count());
     infolog(g_log, "\tMin per-domain:             {}ns", result->match_domains.min_per_domain.count());
     infolog(g_log, "\tMax per-domain:             {}ns", result->match_domains.max_per_domain.count());
@@ -212,10 +212,10 @@ int main(int argc, char **argv) {
 
     result.match_domains.tries = domains.size();
 
-    result.overall.start_rss = ag::dns::sys::current_rss();
+    result.overall.start_rss = static_cast<int>(ag::dns::sys::current_rss());
     TICK(result.overall.start_ts);
 
-    result.load_rules.start_rss = ag::dns::sys::current_rss();
+    result.load_rules.start_rss = static_cast<int>(ag::dns::sys::current_rss());
 
     infolog(g_log, "Loading rules in filter...");
     DnsFilter filter;
@@ -228,7 +228,7 @@ int main(int argc, char **argv) {
         exit(-1);
     }
     TICK(result.load_rules.end_ts);
-    result.load_rules.end_rss = ag::dns::sys::current_rss();
+    result.load_rules.end_rss = static_cast<int>(ag::dns::sys::current_rss());
     if (err_or_warn) {
         warnlog(g_log, "... rules loaded with warnings: {}", err_or_warn->str());
     } else {
@@ -236,15 +236,15 @@ int main(int argc, char **argv) {
     }
 
     infolog(g_log, "Matching domains against rules...");
-    result.match_domains.start_rss = ag::dns::sys::current_rss();
+    result.match_domains.start_rss = static_cast<int>(ag::dns::sys::current_rss());
     apply_filter_to_base(&result, &filter, handle);
-    result.match_domains.end_rss = ag::dns::sys::current_rss();
+    result.match_domains.end_rss = static_cast<int>(ag::dns::sys::current_rss());
     infolog(g_log, "...domains matched");
 
     filter.destroy(handle);
 
     TICK(result.overall.end_ts);
-    result.overall.end_rss = ag::dns::sys::current_rss();
+    result.overall.end_rss = static_cast<int>(ag::dns::sys::current_rss());
 
     report_results(&result);
 }
